requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.0
async-timeout==4.0.3
attrs==23.1.0
bandit==1.7.5
billiard==4.1.0
boto3==1.28.62
botocore==1.31.62
build==0.10.0
CacheControl==0.13.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
crashtest==0.4.1
cryptography==41.0.4
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.7
dulwich==0.21.6
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.4
frozenlist==1.4.0
gitdb==4.0.10
github==1.2.7
GitPython==3.1.37
greenlet==3.0.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
isort==5.12.0
jaraco.classes==3.3.0
jeepney==0.8.0
jmespath==1.0.1
jsonschema==4.17.3
keyring==24.2.0
kombu==5.3.2
mangum==0.17.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
openai==0.28.1
packaging==23.2
pbr==5.11.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pycparser==2.21
pycurl==7.45.2
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyGithub==2.1.1
Pygments==2.16.1
PyJWT==2.8.0
pylint==3.0.1
PyNaCl==1.5.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.4.2
python-dateutil==2.8.2
PyYAML==6.0.1
rapidfuzz==2.15.2
redis==4.6.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.6.0
s3transfer==0.7.0
SecretStorage==3.3.3
shellingham==1.5.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
stevedore==5.1.0
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
trove-classifiers==2023.9.19
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.2.8
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0